requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
boto==2.49.0
boto3==1.28.56
botocore==1.31.56
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.6.1
decorator==5.1.1
elevenlabs==0.2.26
exceptiongroup==1.1.3
executing==1.2.0
Flask==2.3.3
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==20.0.4
idna==3.4
importlib-metadata==6.8.0
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.304
langsmith==0.0.41
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.26.0
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.4.1
pydantic_core==2.10.1
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
replicate==0.13.0
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
stack-data==0.6.2
tenacity==8.2.3
traitlets==5.10.1
twilio==8.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==1.26.16
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.17.0
